USB HID : Make Interrupt endpoints work
- QueueHead: Rename some properties and expose .physQueueHead and
  .nextQH.

- UHCIPipe: Allocate an array of TransferDescriptors and use instead of
  setupTD/statusTD which are specific to control pipes.

- HCD_UHCI: Rename qhStart to controlQH as it is the start of the QHs
  for Control pipes.

- HCD_UHCI: Add addQueueHead() and removeQueueHead() to add new
  QueueHeads into the correct Control or Interrupt QH chain.

- HCD_UHCI: Add dumpAndCheckFrameList() for debugging.

- HID: Interrupt for keyboard now works, although it is polled in a loop.

+++ kernel/devices/usb/usb-endpointdescriptor.swift
@@ -43,7 +43,7 @@ extension USB {
 
         var bmAttributesBits: BitArray8 { BitArray8(bmAttributes) }
 
-        var endpoint: UInt8 { descriptor.bEndpointAddress & 0xf }
+        var endpoint: UInt { UInt(descriptor.bEndpointAddress & 0xf) }
         var direction: TransferDirection { TransferDirection(rawValue: descriptor.bEndpointAddress >> 7)! }
         var transferType: TransferType { TransferType(rawValue: UInt8(bmAttributesBits[0...1]))! }
         var synchronization: SynchronizationType { SynchronizationType(rawValue: UInt8(bmAttributesBits[2...3]))! }
